refactor: api 파일 정리 및 type 정리
