--- src/types/HttpMethod.ts
@@ -1,1 +1,0 @@
-export type HttpMethod = "GET" | "POST" | "PUT" | "DELETE" | "PATCH";
